Paginated types, discord details implemented

+++ src/graphql/resolvers/workspace.ts
@@ -1,5 +1,5 @@
 import {Bot, getModuleRepo, getUserRepo, getWorkspaceRepo, Module, User, Workspace} from '@canalapp/shared/dist/db';
-import {GraphContext} from '../typeDefs';
+import {GraphContext, Paginated} from '../typeDefs';
 
 export interface CreateWorkspaceInput {
   name: string;
@@ -12,27 +12,33 @@ export interface UpdateWorkspaceInput {
 
 export const workspaceResolvers = {
   Query: {
-    async workspaces(parent: void, args: void, context: GraphContext): Promise<Workspace[]> {
-      return await getWorkspaceRepo().find({resourceOwnerId: context.user.id});
+    async workspaces(parent: void, args: void, context: GraphContext): Promise<Paginated<Workspace>> {
+      const workspaces = await getWorkspaceRepo().find({resourceOwnerId: context.user.id});
+      return {
+        totalCount: workspaces.length,
+        nodes: workspaces
+      };
     },
     async workspace(parent: void, args: {id: string}, context: GraphContext): Promise<Workspace | null> {
+      if (!args.id) throw new Error('No id provided');
+
       return await getWorkspaceRepo().findOneIfUserCanRead(args.id, context.user);
     }
   },
   Mutation: {
-    async createWorkspace(parent: void, args: CreateWorkspaceInput, context: GraphContext): Promise<Workspace> {
+    async createWorkspace(parent: void, args: {workspace: CreateWorkspaceInput}, context: GraphContext): Promise<Workspace> {
       return await getWorkspaceRepo().createAndSave({
-        name: args.name,
+        name: args.workspace.name,
         user: context.user,
         personal: false
       });
     },
-    async updateWorkspace(parent: void, args: UpdateWorkspaceInput, context: GraphContext): Promise<Workspace> {
+    async updateWorkspace(parent: void, args: {workspace: UpdateWorkspaceInput}, context: GraphContext): Promise<Workspace> {
       const workspaceRepo = getWorkspaceRepo();
-      const workspace = await workspaceRepo.findOneIfUserCanRead(args.id, context.user);
+      const workspace = await workspaceRepo.findOneIfUserCanRead(args.workspace.id, context.user);
       if (!workspace) throw new Error('Couldn\'t find that workspace');
 
-      if (args.name) workspace.name = args.name;
+      if (args.workspace.name) workspace.name = args.workspace.name;
 
       await workspaceRepo.save(workspace);
       return workspace;
@@ -43,9 +49,12 @@ export const workspaceResolvers = {
       if (!workspace) {
         throw new Error('Couldn\'t find that workspace');
       }
+      if (workspace.isPersonal) {
+        throw new Error('You can\'t delete your personal workspace!');
+      }
 
       await workspaceRepo.remove(workspace);
-      return workspace.id;
+      return args.id;
     }
   },
   Workspace: {
@@ -58,8 +67,12 @@ export const workspaceResolvers = {
     async users() {
       // TODO users
     },
-    async modules(parent: Workspace, args: void, context: GraphContext): Promise<Module[]> {
-      return await getModuleRepo().find({workspaceId: parent.id});
+    async modules(parent: Workspace, args: void, context: GraphContext): Promise<Paginated<Module>> {
+      const modules = await getModuleRepo().find({workspaceId: parent.id});
+      return {
+        totalCount: modules.length,
+        nodes: modules
+      };
     }
   }
 };
